refactor: add Footer component with structured layout, social links, and reusable FooterLink
