+++ src/app/components/footer/index.tsx
@@ -1,0 +1,2 @@
+export { default as FooterLink } from './FooterLink';
+export { default as SocialLink } from './SocialLink';
